✨(front) add create folder and refactor provider
This form creates a folder inside the current one. The ExplorerProvider
has been extraction from Explorer because this one has to be on top of
the ExplorerTree as well to give him access to various contextual data
needed to construct the tree. Some additionnal debug data are also displayed
in the tree to make easier the setup of the tree. I also made some
var renaming for better DX. Also included the current item in the context.