♻️(front) make error handling work on file upload
Using mutateAsync makes the automatic global handling of error
don't work. It forces to catch the error locally each time, while
this is a strange behavior, this is totally assumed by
react query authors. So to bypass this while preversing the
sequential execution of uploading, we need to refactor this that
way.