✨(front) add dropdown to import button
We want to open two kind of file picker, one for files and the other
for folders. Important: at the moment it is not possible to select
multiple folders at once via filer picker.
Also I used a bit of a hackish approach with webkit features and
two input file in order to make the two work, but no other alternative
is given from react drop zone at the moment due to the limitation of
the File System API.