♻️(front) decouple ExplorerContext and Explorer content
We need to adapt the architecture to custom explorer content. For the trash
implementation we do not have a current item, it is simply the trash.
The ExplorerContext could not work without having a current item and it
was also holding the explorer children items, but that's too much for it.
We want to provide a layout with the tree only, the content of the central
part needs to be fully customizable.